Add Monte Carlo cross validation
A corpus is split randomly into a test and training set (based on a
given fraction determining the size of the test set). The tagger trained
on the training set is used to tag the sentences in the test set. This
procedure may be repeated k times. The results are written to a file for
later statistical evaluation.